Removed Redudancies. Added Time adder on game

+++ ADT/point.c
@@ -2,7 +2,7 @@
 #include "point.h"
 
 /* *** CREATOR *** */
-void createPoint(POINT *P, float X, float Y){
+void createPoint(POINT *P, int X, int Y){
 /* Membuat point sembarang*/
     ABSIS(*P) = X;
     ORDINAT(*P) = Y;
@@ -13,7 +13,7 @@ void tulisPoint(POINT P){
 /* Menulis point sekarang dalam format
     (X,Y)
 */
-    printf("(%.2f,%.2f)\n", ABSIS(P), ORDINAT(P));
+    printf("(%d,%d)\n", ABSIS(P), ORDINAT(P));
 }
 
 /* *** OPERASI *** */
